--- src/controllers/helloController.js
@@ -1,4 +1,0 @@
-exports.helloWorld = (req, res) => {
-  console.log('hello world');
-  res.status(200).json({ message: 'hello world' });
-};
