+++ src/init.js
@@ -1,3 +1,7 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
+
 import app from './app';
 
 import { sequelize } from './models';
