--- src/router/helloRouter.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import { helloWorld } from '../controllers/helloController';
-
-const helloRouter = express.Router();
-
-helloRouter.get('/', helloWorld);
-
-export default helloRouter;
